+++ R/prepare_sciCSR.R
@@ -21,11 +21,8 @@ prepare_sciCSR <- function()
     conda_install(envname = 'scicsr', packages = c('scanpy', 'deeptime'), channel = 'conda-forge')
     conda_install(envname = 'scicsr', packages = 'h5py', pip = TRUE, pip_options = "--force-reinstall")
     conda_install(envname = 'scicsr', packages = 'scvelo', pip = TRUE, pip_options = "-U")
-    conda_install(envname = 'scicsr', packages = 'jupyter', channel = 'conda-forge')
-    conda_install(envname = 'scicsr', packages = c('python-igraph', 'louvain'),
-                  pip = TRUE)
-    conda_install(envname = 'scicsr', packages = 'cellrank',
-                  channel = c('conda-forge', 'bioconda'))
+    conda_install(envname = 'scicsr', packages = c('python-igraph', 'louvain'), pip = TRUE)
+    conda_install(envname = 'scicsr', packages = 'cellrank', pip = TRUE)
     conda_install(envname = 'scicsr', packages = 'networkx==2.8.2')
     conda_install(envname = 'scicsr', packages = 'multiprocess')
   }